Add authenticating-proxy to services
This adds the authenticating-proxy app which can run in front of draft-router.
In order to run it against frontend apps they'll also need to be started
up as draft router doesn't have dependencies on them.

This resolves running authenticating-proxy in the default GDS SSO "mock"
mode. In order to run this with GDS SSO in "real" mode there needs to be
co-ordination between this authenticating-proxy and signon to share
oauth data, which was out of the scope of this change.