Rewrite 'compose spec' to check nginx config
Previously we had a spec to check the configured domains for compose
services matched those defined in nginx-proxy-app. This rewrites that
spec to be more comprehensive by enforcing that all 'app' stacks
define the correct config to ensure they work in a browser.

One small loss from the original spec is that we no longer compare the
total set of domains in our apps versus nginx-proxy, which could mean
we end up with nginx-proxy having unused domains configured, which might
cause confusion. We could introduce a spec if that's a problem.