Rename nginx-proxy-app to nginx-proxy
nginx-proxy is not a standard service, and the '-app' defining it
with an 'app' stack is at odds with the expectation that it can be
'run in the browser', as stated in our docs. This removes the notion
of stacks from nginx-proxy as a core dependency, which has the nice
side effect of simplifying the new 'virtual hosts' spec.