Support interactive debugging for Rails apps
This enables interactive terminals on all app stacks, together with a
how-to that explains using 'docker attach' to work with an interactive
inline debugger, which is commonly used with Rails apps. Although it's
possible to get an interactive terminal using 'run', this only works for
the top-level app, whereas this technique also supports using an inline
debugger in background dependencies.

One downside of this change is that it increases the boilerplate for
each of the config files that define a Rails app. Unfortunately this is
a restriction of YAML - it's not possible to define aliases in a central
file, which would have been a great way to consolidate common config.
Making this change does have a beneficial side-effect for long running
web requests, which are now interruptible.