Spike into using volumes to improve SCSS compilation speed
This applies volumes to both govuk_publishing_components and
content-publisher for storing temporary files and node_modules.

Applying this for these apps compiling sassc reduced compilation times
accordingly:

- govuk_publishing_components with a cold cache: ~46s before, ~6s after
- govuk_publishing_components with a warm cache: ~6s before, ~0.5s after
- content-publisher with a cold cache: ~19s before, ~4s after
- content-publisher with a warm cache: ~8s before, ~0.5s after

A big pain point of doing this is that volumes configured by
docker-compose are automatically owned by root [1]. The fastest approach
for me to get a proof of concept together was to switch the user running
docker to root.

[1]: https://github.com/docker/compose/issues/3270