Run Ruby apps as root
This adapts the base Dockerfile that is used for Ruby apps to remove the
creation and running as a user build. This is done so that it can be
easier to create mountable volumes for particular applications. This is
done to provide improved performance on macOS.

In order for apps to be able to continue to run Chrome tests they will
need to have installed govuk_test once this PR is
released: https://github.com/alphagov/govuk_test/pull/31 as this will
allow running chrome in a no-sandbox mode based on the presence of the
`GOVUK_TEST_CHROME_NO_SANDBOX` environment variable. Because of this we
no longer need to set the privileged flag for individual containers.