Use version numbers in database services and mounts
This change has been made to reflect the migration pains developers
experienced when mongo 2.4 was upgraded to 3.6. Users found they
experienced errors trying to use their apps because their mongo mounts
contained databases incompatible with the mongo container.

This change adds the version number to both the mount and service name
so that the version is indicated. When versions of databases change we
can then create a new mount that is only for that version, thereby
starting fresh there. If people wish to maintain data they can always
copy the volume but this wouldn't be automatic.

The approach I took for choosing the fidelity of version numbers was to
look at the options tools like brew provided for installing. For the
(rather old) versions of MySQL, PostgreSQL and Mongo there seemed to be
emphasis on major.minor versions (for example, 9.6 for PostgreSQL) whereas for
newer versions only major versions (for example, 13 for PostgreSQL) -
hopefully once we upgrade past the current major versions we are using
(whenever that day may be) we can flag these entirely at major versions.

I chose not to put versions in the service/mount names of
rabbitmq, redis, and memcache as we don't have any explicit versioning
for them currently. We may want to review this if we are to make these
more representative of what is used on GOV.UK (which is presumably far
older versions).