Mark Whitehall as (basically) compatible
Previously this was based on the ability to do end-to-end publishing,
which is no longer supported [1]. Trying this locally:

- I'm able to render several pages in whitehall-frontend.

- I'm able to draft news stories and do other admin functions.

Running "make" also succeeds. While some (5) of the tests are failing
("reslugging_test" and "govspeak_helper_test"), this is a very small
proportion of the extremely large number of tests in this repo. Since
it's unlikely anyone will need/want to run the entire test suite on
their local machine, it seems reasonable to ignore these few failures,
which may or may not be within the scope of this repo to fix.

[1]: https://github.com/alphagov/govuk-docker/pull/368